Implementacion del metodo buscar Top 5 mascotas mas costosas

+++ src/models/Gato.java
@@ -3,7 +3,7 @@ package models;
 public class Gato extends Mascota{
   
     private RazaGato raza;
-    public Gato(Paises paisOrige, double costo, String nombre, byte edad, RazaGato raza) {
+    public Gato(Paises paisOrige, float costo, String nombre, byte edad, RazaGato raza) {
         super(paisOrige, costo, nombre, edad);
         this.raza = raza;
     }
